svc-face-backup/svc-face/app/app/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
asyncpg==0.29.0
httpx==0.26.0
python-multipart==0.0.6
fal-client==0.4.1
deep-translator==1.11.4
azure-ai-contentsafety==1.0.0
azure-core==1.29.6
azure-storage-blob==12.19.0
Pillow==10.2.0
python-jose[cryptography]==3.3.0
redis==5.0.1
tenacity==8.2.3
PyJWT==2.8.0